@@ -28,12 +28,12 @@
     "locomotive-scroll": "4.1.4",
     "logrocket": "3.0.1",
     "mdx-bundler": "9.0.1",
-    "next": "13.1.1",
+    "next": "^13.4.6",
     "next-contentlayer": "0.2.9",
     "next-themes": "0.2.1",
     "nprogress": "0.2.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "4.6.0",
     "react-locomotive-scroll": "0.2.2",
     "reading-time": "1.5.0",
@@ -61,7 +61,7 @@
     "cross-env": "7.0.3",
     "dedent": "0.7.0",
     "eslint": "8.31.0",
-    "eslint-config-next": "13.1.1",
+    "eslint-config-next": "^13.4.6",
     "eslint-config-prettier": "8.6.0",
     "eslint-plugin-jsx-a11y": "6.6.1",
     "eslint-plugin-prettier": "4.2.1",
