@@ -1,16 +1,18 @@
-import Scroll from 'locomotive-scroll'
-import { useRouter } from 'next/router'
-import { ReactNode, useRef } from 'react'
-import { LocomotiveScrollProvider } from 'react-locomotive-scroll'
-import { ScrollObserver } from './ScrollObserver'
+'use client';
+
+import Scroll from 'locomotive-scroll';
+import { usePathname } from 'next/navigation';
+import { ReactNode, useRef } from 'react';
+import { LocomotiveScrollProvider } from 'react-locomotive-scroll';
+import { ScrollObserver } from './ScrollObserver';
 
 interface Props {
-  children: ReactNode
+  children: ReactNode;
 }
 
 export default function LocomotiveProvider({ children }: Props) {
-  const router = useRouter()
-  const scrollRef = useRef<HTMLDivElement | null>(null)
+  const pathName = usePathname();
+  const scrollRef = useRef<HTMLDivElement | null>(null);
 
   return (
     <LocomotiveScrollProvider
@@ -26,10 +28,10 @@ export default function LocomotiveProvider({ children }: Props) {
         },
       }}
       watch={[]}
-      location={router.asPath}
+      location={pathName}
       containerRef={scrollRef}
       onLocationChange={(scroll: Scroll) => {
-        scroll.scrollTo(0, { duration: 250, disableLerp: true })
+        scroll.scrollTo(0, { duration: 250, disableLerp: true });
       }}
       onUpdate={() => undefined}
     >
@@ -39,5 +41,5 @@ export default function LocomotiveProvider({ children }: Props) {
         </main>
       </ScrollObserver>
     </LocomotiveScrollProvider>
-  )
+  );
 }
