@@ -1,7 +1,7 @@
-import { ReactNode } from 'react'
+import { ReactNode } from 'react';
 
 interface Props {
-  children: ReactNode
+  children: ReactNode;
 }
 
 export default function SectionContainer({ children }: Props) {
@@ -9,5 +9,5 @@ export default function SectionContainer({ children }: Props) {
     <div className="mx-auto max-w-3xl px-4 sm:px-9 xl:max-w-5xl xl:px-0" data-scroll>
       {children}
     </div>
-  )
+  );
 }
