@@ -1,10 +1,13 @@
-import fetcher from 'lib/fetcher'
-import useSWR from 'swr'
-import { NowPlayingSong } from '../../types/Spotify'
-import AnimatedBars from './AnimatedBars'
+'use client';
+
+import fetcher from 'lib/fetcher';
+import useSWR from 'swr';
+import { NowPlayingSong } from '../../types/Spotify';
+import AnimatedBars from './AnimatedBars';
 
 export default function NowPlaying() {
-  const { data: nowPlaying } = useSWR<NowPlayingSong>('/api/now-playing', fetcher)
+  // TODO: fetch this from the server and remove swr
+  const { data: nowPlaying } = useSWR<NowPlayingSong>('/api/now-playing', fetcher);
 
   if (!nowPlaying?.songUrl || !nowPlaying.title || !nowPlaying.artist) {
     return (
@@ -21,7 +24,7 @@ export default function NowPlaying() {
           <p className="text-gray-500 dark:text-gray-300">Spotify</p>
         </div>
       </div>
-    )
+    );
   }
 
   return (
@@ -42,5 +45,5 @@ export default function NowPlaying() {
         </p>
       </div>
     </div>
-  )
+  );
 }
