@@ -1,14 +1,16 @@
-import { motion } from 'framer-motion'
-import { useTheme } from 'next-themes'
-import { useEffect, useState } from 'react'
-import { BsMoonFill, BsSunFill } from 'react-icons/bs'
+'use client';
+
+import { motion } from 'framer-motion';
+import { useTheme } from 'next-themes';
+import { useEffect, useState } from 'react';
+import { BsMoonFill, BsSunFill } from 'react-icons/bs';
 
 const ThemeSwitch = () => {
-  const [mounted, setMounted] = useState(false)
-  const { theme, setTheme, resolvedTheme } = useTheme()
+  const [mounted, setMounted] = useState(false);
+  const { theme, setTheme, resolvedTheme } = useTheme();
 
   // When mounted on client, now we can show the UI
-  useEffect(() => setMounted(true), [])
+  useEffect(() => setMounted(true), []);
 
   return (
     <motion.button
@@ -37,7 +39,7 @@ const ThemeSwitch = () => {
         )}
       </svg>
     </motion.button>
-  )
-}
+  );
+};
 
-export default ThemeSwitch
+export default ThemeSwitch;
